Consolidating logic into a hook phase 1
